split: test_tools_git.py into 5 focused companion files
Split 1102-line test_tools_git.py into:
- test_tools_git.py (retained): merge_worktree core flow
- test_tools_git_classify_fix.py: classify_fix tool tests
- test_tools_git_merge_cleanup.py: cleanup reporting + warnings
- test_tools_git_merge_guards.py: remote tracking + merge commit detection + timing
- test_tools_git_branch.py: create_unique_branch + check_pr_mergeable
- test_tools_git_split.py: structural guard

Registry updates: _test_filter.py, test_test_filter_cascade.py,
arch/test_feature_markers.py, tests/server/CLAUDE.md
