+++ tests/arch/test_feature_markers.py
@@ -28,6 +28,8 @@ _FLEET_CROSS_DIR_FILES: frozenset[Path] = frozenset(
         _TESTS_ROOT / "cli" / "test_fleet_status.py",
         _TESTS_ROOT / "cli" / "test_fleet_list.py",
         _TESTS_ROOT / "server" / "test_tools_dispatch.py",
+        _TESTS_ROOT / "server" / "test_tools_dispatch_validation.py",
+        _TESTS_ROOT / "server" / "test_tools_dispatch_params.py",
         _TESTS_ROOT / "cli" / "test_food_truck_prompt.py",
         _TESTS_ROOT / "cli" / "test_l3_orchestrator_prompt.py",
         _TESTS_ROOT / "cli" / "test_reap.py",
