+++ tests/server/test_tools_kitchen_gate.py
@@ -1,15 +1,12 @@
-"""Tests for tools_kitchen.py: gate mechanics, hook config, recipe packs, refresh, misc."""
+"""Tests for tools_kitchen.py: gate toggle, review gate cleanup, kitchen_id, misc."""
 
 from __future__ import annotations
 
-import asyncio
-import dataclasses
 import json
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from autoskillit.config.settings import QuotaGuardConfig
 from autoskillit.hooks.formatters._fmt_primitives import _HOOK_CONFIG_PATH_COMPONENTS
 from tests.server.conftest import _make_mock_ctx
 
@@ -52,26 +49,6 @@ def test_close_kitchen_disables_gate(tmp_path, monkeypatch):
     mock_ctx.gate.disable.assert_called_once()
 
 
-def test_hook_config_path_is_inside_temp_dir(tmp_path):
-    """_hook_config_path() must resolve to a path inside the project temp directory.
-
-    This invariant ensures .hook_config.json receives automatic gitignore coverage
-    from temp/.gitignore ('*') and can never produce a gitignore gap regardless of
-    whatever entries are present in .gitignore or _AUTOSKILLIT_GITIGNORE_ENTRIES.
-    """
-    from autoskillit.core.io import resolve_temp_dir
-    from autoskillit.server._misc import _hook_config_path
-
-    hook_path = _hook_config_path(tmp_path)
-    temp_dir = resolve_temp_dir(tmp_path, None)
-
-    assert hook_path.is_relative_to(temp_dir), (
-        f"_hook_config_path() returned {hook_path!r} which is NOT inside "
-        f"temp dir {temp_dir!r}. Session-bridge files must live in temp/ "
-        f"to receive automatic gitignore coverage via temp/.gitignore."
-    )
-
-
 # T2c
 def test_close_kitchen_no_file_no_error(tmp_path, monkeypatch):
     """_close_kitchen_handler() doesn't raise when no gate file exists."""
@@ -87,305 +64,6 @@ def test_close_kitchen_no_file_no_error(tmp_path, monkeypatch):
 
     # Gate file was never created — confirm it still does not exist
     assert not tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS).exists()
-
-
-# T-CACHE-1
-@pytest.mark.anyio
-async def test_open_kitchen_primes_quota_cache(tmp_path, monkeypatch):
-    """open_kitchen must call _prime_quota_cache before any run_skill hook fires."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    prime_mock = AsyncMock()
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch("autoskillit.server.tools.tools_kitchen._prime_quota_cache", prime_mock):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-                    await _open_kitchen_handler()
-
-    prime_mock.assert_called_once()
-
-
-# T-CACHE-2
-@pytest.mark.anyio
-async def test_open_kitchen_writes_hook_config_json(tmp_path, monkeypatch):
-    """open_kitchen must write .autoskillit/.hook_config.json with user quota_guard values."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.config.quota_guard.short_window_threshold = 85.0
-    mock_ctx.config.quota_guard.long_window_threshold = 98.0
-    mock_ctx.config.quota_guard.long_window_patterns = ["weekly", "sonnet", "opus"]
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "/custom/path.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx) as mock_get_ctx:
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-                await _open_kitchen_handler()
-
-    assert mock_get_ctx.call_count >= 2, (
-        "_get_ctx must be called in both _open_kitchen_handler and _write_hook_config; "
-        "if call_count < 2 the patch did not cover _write_hook_config's deferred import"
-    )
-    hook_cfg = tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS)
-    assert hook_cfg.exists(), "Hook config file must be written by open_kitchen"
-    data = json.loads(hook_cfg.read_text())
-    assert data["quota_guard"]["cache_max_age"] == 300
-    assert data["quota_guard"]["cache_path"] == "/custom/path.json"
-    assert data["quota_guard"]["buffer_seconds"] == 60
-    # threshold fields are pre-computed into should_block in the cache — not written to hook_config
-    assert "threshold" not in data["quota_guard"]
-    assert "short_window_threshold" not in data["quota_guard"]
-    assert "long_window_threshold" not in data["quota_guard"]
-    assert "long_window_patterns" not in data["quota_guard"]
-    # disabled is always written by _quota_guard_hook_payload
-    # MagicMock.enabled is truthy by default, so disabled must be False
-    assert data["quota_guard"]["disabled"] is False
-    # Confirm kitchen_id rename: hook config must contain 'kitchen_id' (not 'pipeline_id')
-    assert "kitchen_id" in data, (
-        "hook config must contain 'kitchen_id' after rename from 'pipeline_id'"
-    )
-    assert isinstance(data["kitchen_id"], str) and data["kitchen_id"], (
-        "kitchen_id must be a non-empty string (UUID set by _open_kitchen_handler)"
-    )
-
-
-@pytest.mark.parametrize("enabled,expected_disabled", [(True, False), (False, True)])
-@pytest.mark.anyio
-async def test_open_kitchen_bridges_enabled_flag_as_disabled(
-    tmp_path, monkeypatch, enabled, expected_disabled
-):
-    """_write_hook_config() must serialize cfg.enabled as disabled: not cfg.enabled.
-
-    Regression test for bridge completeness: QuotaGuardConfig.enabled must
-    translate to quota_guard.disabled in .hook_config.json so that the hook
-    subprocess can respect the profile-wide opt-out.
-    """
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.config.quota_guard = QuotaGuardConfig(
-        enabled=enabled, cache_max_age=300, cache_path="/p/q.json", buffer_seconds=60
-    )
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-                await _open_kitchen_handler()
-
-    data = json.loads(tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS).read_text())
-    assert data["quota_guard"]["disabled"] is expected_disabled, (
-        f"enabled={enabled} must produce disabled={expected_disabled} in hook config"
-    )
-
-
-# T-CACHE-3
-@pytest.mark.anyio
-async def test_close_kitchen_removes_hook_config_json(tmp_path, monkeypatch):
-    """close_kitchen must remove .autoskillit/.hook_config.json to prevent stale config."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.config.quota_guard.short_window_threshold = 85.0
-    mock_ctx.config.quota_guard.long_window_threshold = 98.0
-    mock_ctx.config.quota_guard.long_window_patterns = ["weekly", "sonnet", "opus"]
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "~/.claude/quota_cache.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import (
-                    _close_kitchen_handler,
-                    _open_kitchen_handler,
-                )
-
-                await _open_kitchen_handler()
-                _close_kitchen_handler()
-
-    hook_cfg = tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS)
-    assert not hook_cfg.exists(), "Hook config must be removed by close_kitchen"
-
-
-# T-CACHE-4
-def test_open_kitchen_handler_is_async():
-    """_open_kitchen_handler must be an async def so it can await _prime_quota_cache."""
-    import inspect
-
-    from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-    assert inspect.iscoroutinefunction(_open_kitchen_handler), (
-        "_open_kitchen_handler must be async"
-    )
-
-
-# REQ-PACK-008: open_kitchen stores active_recipe_packs
-@pytest.mark.anyio
-async def test_open_kitchen_sets_active_recipe_packs(tmp_path, monkeypatch):
-    """After _open_kitchen_handler(), ctx.active_recipe_packs is frozenset()."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-                    await _open_kitchen_handler()
-
-    assert mock_ctx.active_recipe_packs == frozenset()
-
-
-# REQ-PACK-008: close_kitchen clears active_recipe_packs
-def test_close_kitchen_clears_active_recipe_packs(tmp_path, monkeypatch):
-    """After _close_kitchen_handler(), ctx.active_recipe_packs is None."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.active_recipe_packs = frozenset(["research"])
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            from autoskillit.server.tools.tools_kitchen import _close_kitchen_handler
-
-            _close_kitchen_handler()
-
-    assert mock_ctx.active_recipe_packs is None
-
-
-# T-REFRESH-1
-@pytest.mark.anyio
-async def test_open_kitchen_starts_quota_refresh_task(tmp_path, monkeypatch):
-    """After _open_kitchen_handler(), ctx.quota_refresh_task is a running asyncio.Task."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-
-    async def instant_loop(config, *, provider="anthropic"):
-        await asyncio.sleep(0)
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch("autoskillit.server.tools.tools_kitchen._prime_quota_cache", AsyncMock()):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    with patch(
-                        "autoskillit.server.tools.tools_kitchen._quota_refresh_loop", instant_loop
-                    ):
-                        from autoskillit.server.tools.tools_kitchen import _open_kitchen_handler
-
-                        await _open_kitchen_handler()
-
-    assert mock_ctx.quota_refresh_task is not None
-    assert isinstance(mock_ctx.quota_refresh_task, asyncio.Task)
-    mock_ctx.quota_refresh_task.cancel()
-
-
-# T-REFRESH-2
-def test_close_kitchen_cancels_quota_refresh_task(tmp_path, monkeypatch):
-    """_close_kitchen_handler cancels ctx.quota_refresh_task and sets it to None."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_task = MagicMock(spec=asyncio.Task)
-    mock_ctx.quota_refresh_task = mock_task
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            from autoskillit.server.tools.tools_kitchen import _close_kitchen_handler
-
-            _close_kitchen_handler()
-
-    mock_task.cancel.assert_called_once()
-    assert mock_ctx.quota_refresh_task is None
-
-
-# T-REFRESH-3
-def test_tool_context_has_quota_refresh_task_field():
-    """ToolContext must have a quota_refresh_task field defaulting to None."""
-    from autoskillit.pipeline.context import ToolContext
-
-    fields = {f.name: f for f in dataclasses.fields(ToolContext)}
-    assert "quota_refresh_task" in fields
-    assert fields["quota_refresh_task"].default is None
-
-
-# ---------------------------------------------------------------------------
-# Group G — disable_quota_guard tool
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.anyio
-async def test_disable_quota_guard_writes_disabled_flag(tmp_path, monkeypatch):
-    """disable_quota_guard() sets quota_guard.disabled=True in the overlay file."""
-    monkeypatch.chdir(tmp_path)
-    hook_cfg_path = tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS)
-    hook_cfg_path.parent.mkdir(parents=True, exist_ok=True)
-    hook_cfg_path.write_text(
-        json.dumps({"quota_guard": {"cache_path": "/some/path.json", "cache_max_age": 300}})
-    )
-
-    from autoskillit.server import _state
-
-    mock_state_ctx = _make_mock_ctx()
-    mock_state_ctx.project_dir = tmp_path
-    monkeypatch.setattr(_state, "_ctx", mock_state_ctx)
-
-    from autoskillit.server.tools.tools_kitchen import disable_quota_guard
-
-    result_str = await disable_quota_guard()
-    parsed = json.loads(result_str)
-    assert parsed["success"] is True
-
-    overlay_path = tmp_path.joinpath(*_HOOK_CONFIG_OVERLAY_RELPATH)
-    overlay_payload = json.loads(overlay_path.read_text())
-    assert overlay_payload["quota_guard"]["disabled"] is True
-    base_payload = json.loads(hook_cfg_path.read_text())
-    assert "disabled" not in base_payload.get("quota_guard", {})
-
-
-@pytest.mark.anyio
-async def test_disable_quota_guard_denies_headless(tmp_path, monkeypatch):
-    """disable_quota_guard() returns an error when called from a headless session."""
-    monkeypatch.chdir(tmp_path)
-    monkeypatch.setenv("AUTOSKILLIT_HEADLESS", "1")
-
-    from autoskillit.server.tools.tools_kitchen import disable_quota_guard
-
-    result_str = await disable_quota_guard()
-    parsed = json.loads(result_str)
-    assert parsed["success"] is False
-
-
-@pytest.mark.anyio
-async def test_disable_quota_guard_returns_error_when_kitchen_not_open(tmp_path, monkeypatch):
-    """disable_quota_guard() returns an error when the kitchen is not open."""
-    monkeypatch.chdir(tmp_path)
-    from autoskillit.server import _state
-
-    monkeypatch.setattr(_state, "_ctx", None)
-
-    from autoskillit.server.tools.tools_kitchen import disable_quota_guard
-
-    result_str = await disable_quota_guard()
-    parsed = json.loads(result_str)
-    assert parsed["success"] is False
 
 
 # ---------------------------------------------------------------------------
@@ -614,428 +292,3 @@ async def test_open_kitchen_generates_uuid_without_campaign_env(tmp_path, monkey
 
     kitchen_id = mock_ctx.kitchen_id
     assert len(kitchen_id) == 36 and kitchen_id.count("-") == 4
-
-
-# ---------------------------------------------------------------------------
-# Group K — ingredients_only parameter
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_ingredients_only_strips_content(tmp_path, monkeypatch):
-    """open_kitchen(name=X, ingredients_only=True) must omit content from result."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.load_and_validate.return_value = {
-        "content": "name: test\nsteps:\n  s1:\n    tool: run_skill\n",
-        "ingredients_table": (
-            "| Name | Description | Default |\n| task | What to do | (required) |"
-        ),
-        "suggestions": [],
-        "valid": True,
-        "orchestration_rules": "MANDATORY: execute every step",
-        "stop_step_semantics": "stop semantics text",
-        "content_hash": "abc123",
-        "composite_hash": "def456",
-        "recipe_version": "1.0",
-    }
-    mock_ctx.recipes.find.return_value = MagicMock()
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    with patch(
-                        "autoskillit.server.tools.tools_kitchen._apply_triage_gate",
-                        new=AsyncMock(side_effect=lambda r, *a, **kw: r),
-                    ):
-                        from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                        result_json = await open_kitchen(
-                            name="test", ingredients_only=True, ctx=mock_ctx
-                        )
-
-    result = json.loads(result_json)
-    assert result["success"] is True
-    assert result["ingredients_table"] is not None
-    assert "content" not in result
-    assert "orchestration_rules" not in result
-    assert "sous_chef_discipline" not in result
-    assert "stop_step_semantics" not in result
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_ingredients_only_preserves_metadata(tmp_path, monkeypatch):
-    """ingredients_only=True must preserve success, kitchen, version, valid, suggestions."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.load_and_validate.return_value = {
-        "content": "name: test\nsteps:\n  s1:\n    tool: run_skill\n",
-        "ingredients_table": "| Name | Description | Default |",
-        "suggestions": [],
-        "valid": True,
-        "content_hash": "abc123",
-        "composite_hash": "def456",
-        "recipe_version": "1.0",
-    }
-    mock_ctx.recipes.find.return_value = MagicMock()
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    with patch(
-                        "autoskillit.server.tools.tools_kitchen._apply_triage_gate",
-                        new=AsyncMock(side_effect=lambda r, *a, **kw: r),
-                    ):
-                        from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                        result_json = await open_kitchen(
-                            name="test", ingredients_only=True, ctx=mock_ctx
-                        )
-
-    result = json.loads(result_json)
-    assert result["success"] is True
-    assert result["kitchen"] == "open"
-    assert "version" in result
-    assert result["valid"] is True
-    assert result["suggestions"] == []
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_ingredients_only_no_name_ignored(tmp_path, monkeypatch):
-    """ingredients_only=True with name=None should behave like regular no-name open."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result_json = await open_kitchen(ingredients_only=True, ctx=mock_ctx)
-
-    result = json.loads(result_json)
-    assert result["success"] is True
-    assert "content" in result
-    assert result["kitchen"] == "open"
-    assert "version" in result
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_uses_project_dir_for_recipe_lookup(tmp_path, monkeypatch):
-    """open_kitchen must use tool_ctx.project_dir for recipe discovery, not Path.cwd().
-
-    Regression test: when project_dir differs from cwd, recipes must be found
-    in project_dir's .autoskillit/recipes/ directory.
-    """
-    monkeypatch.chdir(tmp_path)  # Ensure cwd != project_dir
-    different_dir = tmp_path / "project_root"
-    different_dir.mkdir()
-
-    # Create a recipe only in project_dir, not in cwd
-    recipe_yaml_text = (
-        "name: test-project-dir-recipe\n"
-        "description: Test recipe for project_dir propagation\n"
-        "autoskillit_version: '0.2.0'\n"
-        "kitchen_rules:\n"
-        "  - Never use native tools\n"
-        "steps:\n"
-        "  stop:\n"
-        "    action: stop\n"
-        "    message: done\n"
-    )
-    recipes_dir = different_dir / ".autoskillit" / "recipes"
-    recipes_dir.mkdir(parents=True)
-    (recipes_dir / "test-project-dir-recipe.yaml").write_text(recipe_yaml_text)
-
-    # Build context with project_dir = different_dir, recipes = RealRecipeRepository
-    from autoskillit.config.settings import AutomationConfig
-    from autoskillit.core.types import DirectInstall
-    from autoskillit.pipeline.audit import DefaultAuditLog
-    from autoskillit.pipeline.context import ToolContext
-    from autoskillit.pipeline.gate import DefaultGateState
-    from autoskillit.pipeline.timings import DefaultTimingLog
-    from autoskillit.pipeline.tokens import DefaultTokenLog
-    from autoskillit.recipe.repository import DefaultRecipeRepository
-
-    real_repo = DefaultRecipeRepository()
-
-    ctx = ToolContext(
-        config=AutomationConfig(features={"fleet": True}),
-        audit=DefaultAuditLog(),
-        token_log=DefaultTokenLog(),
-        timing_log=DefaultTimingLog(),
-        gate=DefaultGateState(enabled=False),
-        plugin_source=DirectInstall(plugin_dir=tmp_path),
-        runner=None,
-        temp_dir=tmp_path / ".autoskillit" / "temp",
-        project_dir=different_dir,
-        recipes=real_repo,
-    )
-
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.gate = ctx.gate
-    mock_ctx.recipes = ctx.recipes
-    mock_ctx.config = ctx.config
-    mock_ctx.project_dir = ctx.project_dir
-    mock_ctx.active_recipe_packs = frozenset()
-    mock_ctx.active_recipe_features = frozenset()
-    mock_ctx.quota_refresh_task = None
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch(
-                    "autoskillit.server.tools.tools_kitchen._apply_triage_gate",
-                    new=AsyncMock(side_effect=lambda r, *a, **kw: r),
-                ):
-                    with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                        with patch(
-                            "autoskillit.server.tools.tools_kitchen._build_hook_diagnostic_warning",
-                            return_value="",
-                        ):
-                            from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                            result_json = await open_kitchen(
-                                name="test-project-dir-recipe", ctx=mock_ctx
-                            )
-
-    result = json.loads(result_json)
-    assert result["success"] is True, (
-        f"open_kitchen failed to find recipe in project_dir={different_dir}. Result: {result}"
-    )
-    assert result.get("kitchen") == "open"
-
-
-def test_get_recipe_uses_project_dir(tmp_path, monkeypatch):
-    """get_recipe must use ctx.project_dir for recipe lookup, not Path.cwd().
-
-    Regression test: when project_dir differs from cwd, get_recipe must
-    find recipes in project_dir's .autoskillit/recipes/ directory.
-    """
-    import yaml
-
-    monkeypatch.chdir(tmp_path)
-    different_dir = tmp_path / "project_root"
-    different_dir.mkdir()
-
-    recipe_yaml = {
-        "name": "test-get-recipe-project-dir",
-        "description": "Test recipe for get_recipe project_dir",
-        "steps": {"s1": {"tool": "run_skill"}},
-    }
-    recipes_dir = different_dir / ".autoskillit" / "recipes"
-    recipes_dir.mkdir(parents=True)
-    (recipes_dir / "test-get-recipe-project-dir.yaml").write_text(yaml.dump(recipe_yaml))
-
-    from autoskillit.recipe.repository import DefaultRecipeRepository
-
-    real_repo = DefaultRecipeRepository()
-
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = different_dir
-    mock_ctx.recipes = real_repo
-
-    with patch("autoskillit.server._state._get_ctx_or_none", return_value=mock_ctx):
-        from autoskillit.server.tools.tools_kitchen import get_recipe
-
-        result = get_recipe("test-get-recipe-project-dir")
-
-    # get_recipe returns raw YAML when found, JSON error dict when not found
-    assert '"error"' not in result, (
-        f"get_recipe failed to find recipe in project_dir={different_dir}. Result: {result}"
-    )
-    assert "test-get-recipe-project-dir" in result, (
-        f"get_recipe did not return the recipe from project_dir. Result: {result}"
-    )
-
-
-# ---------------------------------------------------------------------------
-# _update_hook_config_with_recipe — recipe authorization in hook config
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.anyio
-async def test_update_hook_config_with_recipe_includes_recipe_allows_pr_create(
-    tmp_path, monkeypatch
-):
-    """_update_hook_config_with_recipe adds recipe_allows_pr_create for PR-creating recipes."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.recipe_name = "merge-prs"
-    mock_ctx.kitchen_id = "test-kitchen-id"
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "/p/q.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import (
-                    _open_kitchen_handler,
-                    _update_hook_config_with_recipe,
-                )
-
-                await _open_kitchen_handler()
-                _update_hook_config_with_recipe()
-
-    hook_cfg = tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS)
-    data = json.loads(hook_cfg.read_text())
-    assert data["recipe_allows_pr_create"] is True
-    assert "quota_guard" in data
-    assert "kitchen_id" in data
-
-
-@pytest.mark.anyio
-async def test_update_hook_config_with_recipe_excludes_pr_create_for_non_pr_recipes(
-    tmp_path, monkeypatch
-):
-    """_update_hook_config_with_recipe does NOT set recipe_allows_pr_create for non-PR recipes."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.recipe_name = "smoke-test"
-    mock_ctx.kitchen_id = "test-kitchen-id"
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "/p/q.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import (
-                    _open_kitchen_handler,
-                    _update_hook_config_with_recipe,
-                )
-
-                await _open_kitchen_handler()
-                _update_hook_config_with_recipe()
-
-    hook_cfg = tmp_path.joinpath(*_HOOK_CONFIG_PATH_COMPONENTS)
-    data = json.loads(hook_cfg.read_text())
-    assert "recipe_allows_pr_create" not in data
-
-
-# ---------------------------------------------------------------------------
-# Group H — layered hook config: overlay isolation
-# ---------------------------------------------------------------------------
-
-_HOOK_CONFIG_OVERLAY_RELPATH = (".autoskillit", "temp", ".hook_config_overlay.json")
-
-
-@pytest.mark.anyio
-async def test_disable_quota_guard_survives_write_hook_config(tmp_path, monkeypatch):
-    """disable_quota_guard writes to overlay; subsequent _write_hook_config cannot erase it."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "/some/path.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    from autoskillit.server import _state
-
-    monkeypatch.setattr(_state, "_ctx", mock_ctx)
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                from autoskillit.server.tools.tools_kitchen import (
-                    _open_kitchen_handler,
-                    _write_hook_config,
-                    disable_quota_guard,
-                )
-
-                await _open_kitchen_handler()
-                result_str = await disable_quota_guard()
-                parsed = json.loads(result_str)
-                assert parsed["success"] is True
-
-                _write_hook_config()
-
-    from autoskillit.hooks._hook_settings import read_merged_hook_config
-
-    merged = read_merged_hook_config(tmp_path)
-    assert merged["quota_guard"]["disabled"] is True
-
-
-def test_write_hook_config_does_not_touch_overlay(tmp_path, monkeypatch):
-    """_write_hook_config() only writes the base file and never modifies the overlay."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.kitchen_id = "test-kitchen-id"
-    mock_ctx.config.quota_guard.cache_max_age = 300
-    mock_ctx.config.quota_guard.cache_path = "/p/q.json"
-    mock_ctx.config.quota_guard.buffer_seconds = 60
-
-    overlay_path = tmp_path.joinpath(*_HOOK_CONFIG_OVERLAY_RELPATH)
-    overlay_path.parent.mkdir(parents=True, exist_ok=True)
-    overlay_path.write_text(json.dumps({"quota_guard": {"disabled": True}}))
-    mtime_before = overlay_path.stat().st_mtime
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            from autoskillit.server.tools.tools_kitchen import _write_hook_config
-
-            _write_hook_config()
-
-    assert overlay_path.stat().st_mtime == mtime_before
-    assert json.loads(overlay_path.read_text())["quota_guard"]["disabled"] is True
-
-
-def test_merged_hook_config_overlay_wins():
-    """merge_hook_configs gives overlay precedence over base."""
-    from autoskillit.hooks._hook_settings import merge_hook_configs
-
-    base = {"quota_guard": {"disabled": False, "cache_max_age": 60}, "kitchen_id": "k1"}
-    overlay = {"quota_guard": {"disabled": True}}
-    merged = merge_hook_configs(base, overlay)
-    assert merged["quota_guard"]["disabled"] is True
-    assert merged["quota_guard"]["cache_max_age"] == 60
-    assert merged["kitchen_id"] == "k1"
-
-
-@pytest.mark.anyio
-async def test_close_kitchen_cleans_overlay(tmp_path, monkeypatch):
-    """close_kitchen removes the overlay file, resetting runtime mutations."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.project_dir = tmp_path
-
-    overlay_path = tmp_path.joinpath(*_HOOK_CONFIG_OVERLAY_RELPATH)
-    overlay_path.parent.mkdir(parents=True, exist_ok=True)
-    overlay_path.write_text(json.dumps({"quota_guard": {"disabled": True}}))
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            from autoskillit.server.tools.tools_kitchen import _close_kitchen_handler
-
-            _close_kitchen_handler()
-
-    assert not overlay_path.exists()
