+++ tests/test_test_filter_cascade.py
@@ -121,7 +121,7 @@ _SERVER_FILE_LEVEL_ENTRIES = [
     "test_server_tool_registration.py",
     "test_mcp_overrides.py",
     "test_smoke_pipeline.py",
-    "test_tools_dispatch.py",
+    "test_tools_dispatch_validation.py",
     "test_tools_kitchen_envelope.py",
     "test_service_wrappers.py",
     "test_tools_list_recipes.py",
